+++ src/Dep.Runtime.Serialization.NewtonsoftJson/GlobalSuppressions.cs
@@ -1,0 +1,9 @@
+﻿// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.Design", "BaseMethodViolations:Exception(s) thrown but not declared in base method", Justification = "<挂起>", Scope = "member", Target = "~M:NetowlsStudio.Dep.Runtime.Serializations.Providers.JsonSerializer.InternalDeserialize``1(System.IO.Stream)~``0")]
+[assembly: SuppressMessage("Microsoft.Design", "BaseMethodViolations:Exception(s) thrown but not declared in base method", Justification = "<挂起>", Scope = "member", Target = "~M:NetowlsStudio.Dep.Runtime.Serializations.Providers.JsonSerializer.InternalSerialize``1(``0,System.IO.Stream)")]